matching algo still ongoing

+++ src/main/java/com/pidev/models/EventParticipant.java
@@ -2,6 +2,6 @@ package com.pidev.models;
 
 public interface EventParticipant {
 	long getEventId();
-    int getParticipant();
+    long getParticipant();
 
 }
